fix: Move JSONRPC specific test cases to QuarkusA2AJSONRPCTest
