+++ reference/jsonrpc/src/test/java/io/a2a/server/apps/quarkus/QuarkusA2AJSONRPCTest.java
@@ -2,9 +2,22 @@ package io.a2a.server.apps.quarkus;
 
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.fail;
 import static org.wildfly.common.Assert.assertNotNull;
 import jakarta.ws.rs.core.MediaType;
+
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.Stream;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
 
 import io.a2a.server.apps.common.AbstractA2AServerTest;
 import io.a2a.spec.A2AClientException;
@@ -12,13 +25,22 @@ import io.a2a.spec.InvalidParamsError;
 import io.a2a.spec.InvalidRequestError;
 import io.a2a.spec.JSONParseError;
 import io.a2a.spec.JSONRPCErrorResponse;
+import io.a2a.spec.Message;
+import io.a2a.spec.MessageSendParams;
 import io.a2a.spec.MethodNotFoundError;
+import io.a2a.spec.Part;
+import io.a2a.spec.SendStreamingMessageRequest;
+import io.a2a.spec.SendStreamingMessageResponse;
+import io.a2a.spec.StreamingJSONRPCRequest;
 import io.a2a.spec.Task;
 import io.a2a.spec.TaskQueryParams;
 import io.a2a.spec.TaskState;
+import io.a2a.spec.TextPart;
 import io.a2a.spec.TransportProtocol;
+import io.a2a.util.Utils;
 import io.quarkus.test.junit.QuarkusTest;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
@@ -177,4 +199,102 @@ public class QuarkusA2AJSONRPCTest extends AbstractA2AServerTest {
         }
     }
 
+    @Test
+    public void testStreamingMethodWithAcceptHeader() throws Exception {
+        testSendStreamingMessage(MediaType.SERVER_SENT_EVENTS);
+    }
+
+    @Test
+    public void testSendMessageStreamNewMessageSuccess() throws Exception {
+        testSendStreamingMessage(null);
+    }
+
+    private void testSendStreamingMessage(String mediaType) throws Exception {
+        Message message = new Message.Builder(MESSAGE)
+                .taskId(MINIMAL_TASK.getId())
+                .contextId(MINIMAL_TASK.getContextId())
+                .build();
+        SendStreamingMessageRequest request = new SendStreamingMessageRequest(
+                "1", new MessageSendParams(message, null, null));
+
+        CompletableFuture<HttpResponse<Stream<String>>> responseFuture = initialiseStreamingRequest(request, mediaType);
+
+        CountDownLatch latch = new CountDownLatch(1);
+        AtomicReference<Throwable> errorRef = new AtomicReference<>();
+
+        responseFuture.thenAccept(response -> {
+            if (response.statusCode() != 200) {
+                //errorRef.set(new IllegalStateException("Status code was " + response.statusCode()));
+                throw new IllegalStateException("Status code was " + response.statusCode());
+            }
+            response.body().forEach(line -> {
+                try {
+                    SendStreamingMessageResponse jsonResponse = extractJsonResponseFromSseLine(line);
+                    if (jsonResponse != null) {
+                        assertNull(jsonResponse.getError());
+                        Message messageResponse =  (Message) jsonResponse.getResult();
+                        assertEquals(MESSAGE.getMessageId(), messageResponse.getMessageId());
+                        assertEquals(MESSAGE.getRole(), messageResponse.getRole());
+                        Part<?> part = messageResponse.getParts().get(0);
+                        assertEquals(Part.Kind.TEXT, part.getKind());
+                        assertEquals("test message", ((TextPart) part).getText());
+                        latch.countDown();
+                    }
+                } catch (JsonProcessingException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+        }).exceptionally(t -> {
+            if (!isStreamClosedError(t)) {
+                errorRef.set(t);
+            }
+            latch.countDown();
+            return null;
+        });
+
+
+        boolean dataRead = latch.await(20, TimeUnit.SECONDS);
+        Assertions.assertTrue(dataRead);
+        Assertions.assertNull(errorRef.get());
+
+    }
+
+    private CompletableFuture<HttpResponse<Stream<String>>> initialiseStreamingRequest(
+            StreamingJSONRPCRequest<?> request, String mediaType) throws Exception {
+
+        // Create the client
+        HttpClient client = HttpClient.newBuilder()
+                .version(HttpClient.Version.HTTP_2)
+                .build();
+
+        // Create the request
+        HttpRequest.Builder builder = HttpRequest.newBuilder()
+                .uri(URI.create("http://localhost:" + serverPort + "/"))
+                .POST(HttpRequest.BodyPublishers.ofString(Utils.OBJECT_MAPPER.writeValueAsString(request)))
+                .header("Content-Type", APPLICATION_JSON);
+        if (mediaType != null) {
+            builder.header("Accept", mediaType);
+        }
+        HttpRequest httpRequest = builder.build();
+
+
+        // Send request async and return the CompletableFuture
+        return client.sendAsync(httpRequest, HttpResponse.BodyHandlers.ofLines());
+    }
+
+    private SendStreamingMessageResponse extractJsonResponseFromSseLine(String line) throws JsonProcessingException {
+        line = extractSseData(line);
+        if (line != null) {
+            return Utils.OBJECT_MAPPER.readValue(line, SendStreamingMessageResponse.class);
+        }
+        return null;
+    }
+
+    private static String extractSseData(String line) {
+        if (line.startsWith("data:")) {
+            line =  line.substring(5).trim();
+            return line;
+        }
+        return null;
+    }
 }
